Skip lint check for dynamic react-native version (#26222)
Summary:
The `dependencies` section in the `build.gradle` Android template contains this line:

    implementation "com.facebook.react:react-native:+"  // From node_modules

It causes the following Gradle Lint warning:

> Avoid using '+' in version numbers, can lead to unpredictable or unrepeatable builds

In this case, as the `// From node_modules` comment suggests, the version is _not_ determined by Gradle but by the version specified in `package.json` - Using "+" is completely fine and intentional.

Therefore it can safely be ignored, which is what the added `//noinspection` comment does.

## Changelog

[Android] [Fixed] - Skip lint check for dynamic react-native version
Pull Request resolved: https://github.com/facebook/react-native/pull/26222

Test Plan: Projects generated with the new template no longer cause the warning anymore.

Differential Revision: D17091476

Pulled By: osdnk

fbshipit-source-id: 910b4ecf22ccd11cade9427af68087ffcce8bc8e