Fabric: Making `EventBeat::beat` protected member
Summary:
The documentation for the method says that it must be called from a subclass, but the methods wasn't marked as protected.

Calling `beat` directly will bypass custom logic implemented in subclasses (because it's not a virtual function (by-design)). That's practically what happened before and what the previous diffs in the stack fix.

Changelog: [Internal]

Reviewed By: JoshuaGross

Differential Revision: D18339594

fbshipit-source-id: 3d8c53b69e1b72d447714dd91c9392bdfb05b9ae