Integrate Picker into Fabric on Android
Summary:
In this diff we integrate the Picker component on Android in Fabric. Since both AndroidDropdownPickerNativeComponent and AndroidDialogPickerNativeComponent use `style?: ?TextStyleProp`, which is not supported by the JS codegen, I had to handwrite the C++ files and check them in.

The component is not fully functional yet because `setNativeProps` is not supported in Fabric. I will fix this in the next diff.

Changelog:
[Android] [Added] - Integrate Picker into Fabric on Android

Reviewed By: mdvacca

Differential Revision: D17954435

fbshipit-source-id: 6b7b029ab0c84c27a48c7dddd66878c9dea324bf