Ensure that TextAttribute prop parsing in BaseTextProps uses proper defaults
Summary:
Motivation: in Marketplace, there's a TextInput in JS that sends { lineHeight: null } to C++. In Paper this was correctly handled as a default value, which just causes that property to not be set. In C++ in Fabric, it was being parsed incorrectly as 0 (because of incorrect defaults being passed into `convertRawProp`), causing the text to not be visible at all. The solution is to make sure that its default value is NaN, which is handled correctly, and causes the Text to render again.

Changelog: [Internal]

Reviewed By: shergin

Differential Revision: D19128049

fbshipit-source-id: fe985428f3ed8b90d56cfa387fbc2d1476d19d36