create-react-class is no longer used by RN, so it can be removed (#27861)
Summary:
Now that react-native has got rid of all the uses of create-react-class, it doesn't need to be listed as a dependency.

## Changelog

[General] [Removed] - Removed create-react-class from dependencies
Pull Request resolved: https://github.com/facebook/react-native/pull/27861

Test Plan: Will hit build break when building bundle if its still needed.  Other risks are if community packages are relying on create-react-class in their code, but not declaring it in their dependencies. -- But those packages should fix their dependencies.

Differential Revision: D19568065

Pulled By: cpojer

fbshipit-source-id: be26c3376aaebfa829def12983b04349ba9907ae